+++ der_floh-tick_upgrade_mod/scenes/mod_tick_upgrade_tab.gd
@@ -44,21 +44,6 @@ func _ready() -> void:
 	preserve_bool.check_button.button_pressed = mod_main.get_preserve_duration()
 	preserve_bool.new_value.connect(_on_preserve_duration_changed)
 
-	# --- Section: Debug ---
-	var debug_header := RichTextLabel.new()
-	debug_header.bbcode_enabled = true
-	debug_header.text = "[b]Debug[/b]"
-	debug_header.fit_content = true
-	debug_header.custom_minimum_size = Vector2(0, 28)
-	vbox.add_child(debug_header)
-
-	var debug_bool: SettingBool = BOOL_SCENE.instantiate()
-	debug_bool.default = false
-	vbox.add_child(debug_bool)
-	debug_bool.setting_label.text = "Debug Logging (prints tick speed events to modloader.log)"
-	debug_bool.check_button.button_pressed = mod_main.get_debug_logging()
-	debug_bool.new_value.connect(_on_debug_logging_changed)
-
 
 func _on_always_tick_distant_changed(value: bool) -> void:
 	var mod_main = load("res://mods-unpacked/der_floh-tick_upgrade_mod/mod_main.gd")
@@ -72,8 +57,3 @@ func _on_preserve_duration_changed(value: bool) -> void:
 	ModLoaderLog.info("Preserve duration set to %s" % str(value), LOG_NAME)
 
 
-func _on_debug_logging_changed(value: bool) -> void:
-	var mod_main = load("res://mods-unpacked/der_floh-tick_upgrade_mod/mod_main.gd")
-	mod_main.set_debug_logging(value)
-	mod_main._debug = value
-	ModLoaderLog.info("Debug logging set to %s" % str(value), LOG_NAME)