+++ der_floh-ore_value_mod/changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.0.3] – 2026-06-09
+
+### Removed
+
+- Debug Logging toggle from settings tab; toggle `debug_logging` via the config JSON file instead
+
 ## [1.0.2] – 2026-04-27
 
 ### Added