+++ der_floh-passive_drop_mod/changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.2.4] – 2026-06-09
+
+### Removed
+
+- Debug Logging toggle from settings tab; toggle `debug_logging` via the config JSON file instead
+
 ## [2.2.3] – 2026-04-27
 
 ### Added