+++ der_floh-tick_upgrade_mod/changelog.md
@@ -1,3 +1,9 @@
+## [1.2.3] – 2026-06-09
+
+### Removed
+
+- Debug Logging toggle from settings tab; toggle `debug_logging` via the config JSON file instead
+
 ## [1.2.2] – 2026-04-27
 
 ### Changed